:truck: 점프 로직 JumpComponent 로 이동
