+++ Runelight-Smash-2021/Assets/Scripts/Character/ControllableUnit.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.Events;
 
 public class ControllableUnit : PhysicsUnit
 {
@@ -9,27 +8,12 @@ public class ControllableUnit : PhysicsUnit
     protected CapsuleCollider2D capsule;
     protected JoystickComponent joystickComponent;
     protected AirMovementComponent airMovementComponent;
+    protected JumpComponent jumpComponent;
 
     // Input Variables
     protected bool isInputEnabled = true;
 
-    protected JumpEventType jumpEventType = JumpEventType.None;
-
     private Vector2 slopeStickPosition;
-
-    // Jump Variables
-    protected bool isJumpSquatting;
-    protected bool isJumping;
-    public float shortHopHeight = 1.5f;
-    public float fullHopHeight = 3.0f;
-    public float doubleJumpHeight = 3.0f;
-    public int maxDoubleJumpCount = 1;
-    protected int doubleJumpLeft;
-    public bool canJumpChangeDirection = true;
-    public float minSlopeJumpAngle = 15.0f;
-
-    // Events
-    public UnityEvent onJumpEvent = new UnityEvent();
 
     protected override void Start()
     {
@@ -37,110 +21,18 @@ public class ControllableUnit : PhysicsUnit
         capsule = GetComponent<CapsuleCollider2D>();
         joystickComponent = GetComponent<JoystickComponent>();
         airMovementComponent = GetComponent<AirMovementComponent>();
-        doubleJumpLeft = maxDoubleJumpCount;
-        slopeComponent.onLandEvent.AddListener(OnLand);
+        jumpComponent = GetComponent<JumpComponent>();
     }
 
     protected override void Tick()
     {
-        ApplyControls();
         base.Tick();
         StickToSlope();
     }
 
-    private void ApplyControls()
-    {
-        ApplyJumpMovement();
-    }
-
-    protected virtual void ApplyJumpMovement()
-    {
-        switch (jumpEventType)
-        {
-            case JumpEventType.Start:
-                isJumpSquatting = true;
-                break;
-            case JumpEventType.ShortHop:
-                if (!slopeComponent.isGrounded)
-                {
-                    break;
-                }
-                ShortHop();
-                break;
-            case JumpEventType.FullHop:
-                if (!slopeComponent.isGrounded)
-                {
-                    break;
-                }
-                FullHop();
-                break;
-            case JumpEventType.DoubleJump:
-                if (slopeComponent.isGrounded)
-                {
-                    break;
-                }
-                DoubleJump();
-                break;
-        }
-        jumpEventType = JumpEventType.None;
-    }
-
-    private void ShortHop()
-    {
-        Jump(shortHopHeight);
-    }
-
-    private void FullHop()
-    {
-        Jump(fullHopHeight);
-    }
-
-    private void DoubleJump()
-    {
-        if (doubleJumpLeft <= 0)
-        {
-            return;
-        }
-        doubleJumpLeft -= 1;
-        Jump(doubleJumpHeight);
-    }
-
-    private void Jump(float jumpHeight)
-    {
-        float targetJumpSpeed = joystickComponent.joystick.x > 0 ? Mathf.Max(joystickComponent.joystick.x * airMovementComponent.maxAirSpeed, velocityComponent.velocity.x) : Mathf.Min(joystickComponent.joystick.x * airMovementComponent.maxAirSpeed, velocityComponent.velocity.x);
-        float jumpSpeed = canJumpChangeDirection ? targetJumpSpeed : velocityComponent.velocity.x;
-        Vector2 jumpVelocity = new Vector2(jumpSpeed, Mathf.Sqrt(2.0f * gravityComponent.gravity * jumpHeight));
-        float jumpAngle = Vector2.Angle(Vector2.right, jumpVelocity);
-        float slopeAngle = Mathf.Sign(velocityComponent.velocity.y) * slopeComponent.centerSlopeAngle;
-
-        // Make sure jump velocityComponent.velocity is away from the grounds
-        if (jumpVelocity.x < 0.0f && 180 - jumpAngle < slopeAngle + minSlopeJumpAngle)
-        {
-            float rotateAngle = Mathf.Deg2Rad * -((slopeAngle - (180 - jumpAngle)) + minSlopeJumpAngle);
-            float cos = Mathf.Cos(rotateAngle);
-            float sin = Mathf.Sin(rotateAngle);
-
-            jumpVelocity.x = jumpVelocity.x * cos - jumpVelocity.y * sin;
-        }
-        else if (jumpVelocity.x > 0.0f && jumpAngle < slopeAngle + minSlopeJumpAngle)
-        {
-            float rotateAngle = Mathf.Deg2Rad * (slopeAngle - jumpAngle + minSlopeJumpAngle);
-            float cos = Mathf.Cos(rotateAngle);
-            float sin = Mathf.Sin(rotateAngle);
-
-            jumpVelocity.x = jumpVelocity.x * cos - jumpVelocity.y * sin;
-        }
-
-        velocityComponent.velocity = jumpVelocity;
-        isJumpSquatting = false;
-        isJumping = true;
-        slopeComponent.isGrounded = false;
-        onJumpEvent.Invoke();
-    }
-
     private void StickToSlope()
     {
-        if (isJumping || !slopeComponent.isGrounded || !slopeComponent.canWalkOnSlope)
+        if (jumpComponent.isJumping || !slopeComponent.isGrounded || !slopeComponent.canWalkOnSlope)
         {
             return;
         }
@@ -204,23 +96,6 @@ public class ControllableUnit : PhysicsUnit
             UnityEditor.Handles.color = Color.green;
             UnityEditor.Handles.DrawWireDisc(slopeStickPosition, Vector3.forward, capsule.size.x / 2);
         }
-    }
-
-    protected void OnLand()
-    {
-        doubleJumpLeft = maxDoubleJumpCount;
-        isJumping = false;
-    }
-
-
-
-    public void SetJumpInput(JumpEventType input)
-    {
-        if (!isInputEnabled)
-        {
-            return;
-        }
-        jumpEventType = input;
     }
 
     public void EnableInput()
