+++ Runelight-Smash-2021/Assets/Scripts/Component/JumpComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b62cf03d15c02c8418d2d840bf853d55
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 190
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
